mib-conf.h: add keys' names for ROI suppport
This commit begin the implementation of ROI (Region of Interest) in
VIVOE. I have decided that the user could defined default values for ROI
in the configuration files. This commit creates the names of the keys
under which we should be able to found those values.
According to the keys found in the configuration file, the ROI for a
Service Provider will be:
_non-scalable
_decimated
_interpolated

For now, only ROIs in Service Provider part will be implemented.