roi.c: add function to handle roi
I add many functions into roi.c. Those functions will assume that the
videoFormat_entry to which thay are related has already be filled once.
Indeed, to work properly we need to have the stream resolution, and I
thought it would be clever to retrieve this information from the
videoFormat_entry. So the functions added in roi.c aims to be called
after the first detection has been performed.
These function will get the ROI parameters from the configuration files,
update the videoFormat_entry's corresponding parameters, adapt the
pipeline properly ( by adding a videocrop element for non-scalable ROI,
and a videoscale element for a decimated or interpolated ROI). It
remains to create the function that will be used to update the pipeline,
in case the user want to set up a ROI for a playing stream.