Implement ROI for RAW video and SP
This commit implements the ROI only for a device that will use RAW
video and that is a Service Provider. Indeed, the dilema here is that
the GstElement videocrop and videosale used to perform the ROI can only
accept RAW videos as src and sink pads' caps. This have been tested with
a non-scalable ROI and it worked. The ROI on the SU's side have been
implemented yet.