roi.c: add function to create ROI from videoChannel
This function aims to update the channel's pipeline when a ROI is set in
the MIB. If the pipeline already contained the element videocrop that we
use to create the pipeline, then its parameters are only updated.
However, if it did not then it is added into the pipeline.
In the case where the video is MPEG-4 or JPEG2000, a decoder and an
encoder are also added as we can only perform cropping on RAW videos.