channelTable.c:remove range checking for ROI params
When I implemented the MIB's member channelTalbe I added a range
checking for the channel's resolution and the channel's roi parameter
values. But this is useless now that I have also implemented the
possibility to use no-filter for the resolution so that the user can use
all resolutions and not only the resolutions restricted by VIVOE.
Furhtermore the parameters HorziRes and VertReas reprensent the
resolution of the ROI when the channel is a ROI, so it cannot be limited
to the minimum and maximum value of the Resolution.