channelTable.c: update VF_entry on ROI changes
I add in channelTable.c a function to update the channel's corresponding
entry in videoFormat Table in case there is a modification of the
channel's ROI parameters. In that case, the corresponding parameters in
videoFormatEntry are also updated.
I also add a function to handle the SET requests performed on ROI
parameters.