replace sink by udp_elem in stream_data
This is more convenient as this is normally used to update video channel
when we change its associated videoFormat. We also need to update the
caps of the updsrc element of SU in case of changes in the SDP so it is
convenient to store this udpsrc element in the stream_data structure.