Make ROI resolution fit in frame resolution
If the difference between th RoiOriginTop and RoiVertRes is greater than
the total frame size of the video then the Roi will be decimated to fit
in the frame. As an example if the vertical frame resolution is 480, and
if the user desires a RoiTopOrigin of 100, with an RoiSize of 480 too,
then the Roi will be -100 under teh frame size. So it will be decimated
to fit into the frame resolution and teh Roi resolution will be 380.