channelTable.c: Make test for ROI in SET_RESERVE_2
The tests were all performed in a function called hanler_roi_requests(),
but there is a special case to tst and accept or refuse the value given
in the request for a particuler MIB member. This case is called
SET_RESERVE_2 in net-snmp. I used this section to test if an MIB "extent
object" can be set i.e. if the corresponding Top and Left origin member
of the MIB are also set.