channelTable.c: update SU's LIB on ROI changes
This was not done: when a NEW SDP file with new caps did arrive the SU's
channel MIB was not updated. This have been correted now it is updated.
However, the way to update channel is not right for now: it is performed
using the function update_channelTable_entry_roi_from_caps () that takes
the new arrived caps in parameter, but this will not work with MPEG-4
videos as the field that represent the frame's resolution are not
explicitely set, but are implicit.
We need to call this function after the performing the typedind after
parsing the MPEG-4 video.