mib-config.: add function to detect vivoe-roi element
We have decided that the ROI would be implemented as follows:
_ We cannot use our own -genral purpose- scaling element videoscale
because scaling is too CPU consuming. Our guiding principle has always
been to let CPU consuming part being performed by the user in the
configuration file (as for encoding/decoding). So we would like to do
the same for videoscaling.
So we will create a vivoe-roi element that would be placed after the
user scaling element, and that will handle the roi caps parameters for
us.

So I created a function in mib-config.c that should parse the
configuration file to see if the current channel being constructed is or
not a ROI. If so it will it will be detected.