mibParameter.h: save pipeline before and after vivoe-roi in roi_datas
This can be performed in mib-config.c. This would prevent from accessing
three times to the configuration file, when it can be accessed only
once. When the initiation of roi_datas is performed we will parse the
gst_source command line so we willl save the pipeline before and after
the "vivoe-roi" element.
Then in stream.c: when created the source we will get the pipeline
before vivoe-roi.
In roi.c: we will add the pipeline after vivoe-roi, once we add the
videocrop or capsfilter to after the source.