channelTable.c: undo videoFormat on failure
When a failure occurs on a ROI changes, the corresponding modified
videoFormat_entry needs to be reset. This is now performed in the UNDO
statement. Furthermore, the videoFormatTable_entry corresponding to the
channel is now accessed only once in channelTable_handler(), unlike
before, where it was accessed in each case. I know this is not optimal,
because in some cases, the videoFormat will never be modified. But when
it is a ROI modification, this will be way more efficient.