roi.c: Make scalable ROI work
The trick is to return TRUE when the ROI parameters are modified, but not
the pipeline. As an example if a user wants to modify the 4 parameters
of the scalable ROI. He needs to start by modifying the TOP and LEFT
parameters, but in that case the height and width values will be
negative as extent parameters are still 0. Then he will modify the
extentRight or extentBotttom parameters. Once again, hight or width will
remain negative. But if the function
update_pipeline_SP_non_scalable_roi_changes return FALSE, the values set
by the user will not be taken into account. So we need to return TRUE.
So it is the user's repsonsability to set ROI parameters, in a way no
error such as a negative resolution (or null) occurs.