roi.c: modify the way to interprate ROI
The ROI are not what was implemented. Indeed, I thought that the ROI
resolution for scalable ROI was the sustraction of Extent - Origin
parameter but that is not it.
Indeed, the Resolution should always be the value of the Res channel
Parameters. But, if the Roi Extent - Origin is smaller or greater than the
resolution, then the video should be scale to the Resolution specified in
the channel Resolution parameters.
It should be decimated if the Resolution of the channel is less than
extent-origin and interpolated if it is greater.

So the implementation need to change. Actually every video are cropped.
So we always need a videocrop element in our pipeline. What is different
is that for scalable ROI we will need a scaling element, for scalable
ROI that should be inserted after the Cropping element. The parameters
in Origin and Extent will defined the value for the crop, then the
parameters in channelHorzRes and channelVertRes will be used for the
scaling element.

In order to do so, I decided to place the vivoe-roi element before the
scaling element, si the element right after the vivoe-roi element should
be the scaling element. In roi.c, I created a function that build a bin
from the element after "vivoe-roi" element, and before the next "!".
This function also modify the content of gst_cmdline_after_roi so it
contains what is after the scaling element in gst_source
because we will place a capsfilter after the scaling element.