roi.c: make scalable ROI work for MPEG-4 videos
The ROI did not work with ROI video before. The reason was that the
resolution of a MPEG-4 video is contained in the parameter "config".
This parameter though needs to be re-generated each time the ROI
parameters change. To do so a new typefinding function needed to be
wrote. This function unlink the input element given in parameter and
the sink element also given in parameter, insert a typefind element
between those two, run the typefinding, and returns the caps between the
two elements.

Doing so we are able to get the new "config" property of the video to
send within our SDP message.