gstvideocrop: Used videocrop source code
I realize that it was not posssible to use an Gsteamer's element inside
the function transform_frame_ip() because this function should be used to
perform the transformation directly on the frame received. One of the
solution I thought was possible was to link src pad of viveocrop plugin
to src pad of the real videocrop element, and link sink pad of both
elements together. It did not work.
I talked to Vincent Dehors, a co-worker of mine, and he advises me to
either write the plugin from scratch, or adapt videocrop source code to
fit my need. After a bit of thinking I choose to adat videocrop plugin
source code. And here the commit. This commit allows to build a local
element named vivoecrop that can be used just as the vidoecrop element
with parameters such as top, bottom, left and right.