adapt code to new vivoecrop element
I have adapted my deprecated code to use the new plugin I have made
vivoecrop.
I have adapted stream.c to parse entirely the bin even if it is a roi,
which was not done before.
I have also adapted some functions from roi.c file. Basically now it only
specifies if the channel is a roi or not, and set the videoFormatType.