gstvivoecrop.c: add new "private" property
I add a property videoFormatIndex to the element vivoecrop. But this
porperty is "private" in the way that it can only be set by the
application's code. The user cannot put in the gst_sourcde cmdline:
gst_source= .... ! vivoecrop videoformatindex=2 ! .....
In order to do so, not only we must add a new member in the structure
GstVivoeCrop but also we must implement our own public setter and
getter. This is what I did.