gstvivoecrop.c: set props with values get from MIB
The plugin vivoecrop does exctaly the same job as the plugin videocrop,
but it set the properties top, left, right and bottom with values it
extracts from the MIB instead of values given by the user. So I add a
function gst_vivoe_crop_update () that get the values entered by the
user in the MIB to update the properties top, left, right, bottom.