remove roi_data structure from code
This is not usefull anymore. Indeed, it has been replaced by a member
roi_scalable in the structure videoFormatTable_entry. The concept of
roi_data is not usefull anymore because we do not parse the
configuration to look for the vivoecrop element name. Instead we
iterate through the pipeline to find the element vivoecrop.