gstvivoecrop.c: remove properties from element
The element vivoecrop has been built on the code of videocrop element.
But the element videocrop element has propoerties that are not usefull
for the element viveocro. I mean they are should remain in the
surtucture GstVivoeCrop so we can actually crop the video, but the
should only be defined as members of the structure and not as properties
of the element. So I removed the part of the code as the getter and the
setter methode for those properties.