vivoecaps.c: create new gstreamer element
Just as I did with vivoecrop element I have also created a vivoecaps
element.
For the ROI, rather than parsing the configuration file in order to look
for the gstreamer 's scaling element used by the user, and then add a
vivoecrop element before it, and a capsfilter after it, we have decided
to create two Gstreamer's plugin so they can be directly used in a
gst_parse function. Using that method we do not have to parsed the
configuration twice.