videoFormatTable: remove roi_scalable member
I have removed the boolean member roi_scalable from the structure
videoFormatTable_entry as it is deprecated. Indeed, now that the element
vivoecaps has been created, there is a simple way to know if a ROI is
indeed scalable. To do so, when the element vivoecaps is detected in the
pipeline, in the function update_pipeline_on_roi_changes() I set a
boolean variable to TRUE that I passed in parameter to the function
gst_vivoe_crop_update() so the plugin vivoecrop can correctly crop the
video if the ROI is scalable.