roi.c: do not force caps on vivoecaps
When the pipeline of gst_source (in a ROI case) is started. I was
forcing the caps after the vivoecaps to be the same as before the
vivoecaps. This is useless. Resolution won't change, and worst in some
cases it ends up with Internal Data Flow Error.
So I simply removed it.