pipeline.c: add function create_branch_in_pipeline
Both scalable and non-scalable roi are now implemented. The only roi
features remaining is to implement them for MPEG-4 video.
In order to do so I will implement a "tee" element in the pipeline.
Indeed, the pipeline will be forked into to branches after the RTP
payloader element in SP's pipeline. One branch will be the actual SP
pipeline, the same as the one implemented for now. The other branch will
only contain a typefind and maybe a fakesink to be able to run typefind
in the pipeline after RTP payloader. This will be used to re-compute
rtp_data and especially the config element for MPEG-4 video.