detect.c: create typefind element especially for roi
This function added in detect.c type_detection_element_for_roi() will
be use to create a typefind element especially for the ROI management.
This element, is not only a reguler typefind element because it has the
a call_back connected to its signal have type. The call_back is the same
as the one we are using for other typefind elements.