pipeline.c: create branch in pipeline only for roi
The branches are created to add a typefind element to use for MPEG4 videos
to re-compute the config parameter. But main of the channels will not be
ROI. These channels will not need two branches in their pipeline. So now,
the branches are created only if the channel is a ROI.