gstvivoecrop.c: remove member videoformatIndex
This was used to associate to each element vivoecrop an entry in the
VideoFormatTable. The goal was to iterate through the list of VF_entry
only on ROI changes (when vivoecrop parameters need to be updated). But
as for scalable or or for MPEG-4 video we also need to iterate the list
of VF_entry, we have the choice:
_ To iterate through the list three two times: one for the cropping
element and one for the vivoecaps or the MPEG-4 video.
_ or only one time, but when the ROI is not scalabe and the video is not
MPEG-4 encoding.
So I prefer to always iterates through the list one rather than
iterate sometimes once, somtimes twice.