mib-config.c: keep configuration file open while init sources
The configuration file can be openned each time we init a source or only
once while sources are initiated and then close. I prefer the second
option which is, I think, more optimal. Before configuration file was
openned once to init the MIB and the redirection_data, and then each
time we were launching a source. Now it is openned once for all that. To
do so I have created a global GKayfile in mibParameters.h to store the
openned configuration file.Once the streams sources have been launched,
it is closed in open vivoe deamon.
But is remains openned each time A SU is created. The reason is that
Service User are created asynchronously, while SP are all created at the
same time.
So the function init_sources_from_conf() and init_sink_from_conf() are
then different in the way that the first one uses the global GKeyFile,
while the second one open the vivoe-mib.conf file each time it is
called.