roi.c:Use new caps instead of copy from input caps
In my previous commit I said that we should be using a copy from the
received input caps. This is wrong. We should absolutly not. Indeed,
input caps can be MPEG4, J2K or RAW. But after a scaling element we
should only have RAW video. The good practice is to built caps of RAW
video as it is the only kind of media we can handle there.