roi.c: change functions name
In order to later implement the ROI for SU, I am preparing the ground!
So I rename set_roi_parameters() to get_roi_values_from_conf() that
stills get the ROI parameter for SP in conf file. And I create a
function that set the values to vivoecrop and vivoecaps outside the
handle_roi() function.