make ROI work for SU in default StartUp mode
I have implemented the ROI for Service User in default Start Up mode. As
a SU's channel does not have an associated videoFormat. When reading the
parameter entered for the channel in configuration file, we copy the ROI
params into the video_stream_info which is a fake videoFormat_entry.
It remains to implement the ROI update for SU. I think a good way to do
so is to create a fake videoFormat_entry copy the ROI params of the
channel into it, and gives this entry as a parameter to the vivoecrop
and vivoecaps element. Another solution is to create in the code of the
vivoecrop and vivoecaps element the same function that work with
videoFormat_entry, but for channelTable_entry.