Finalize ROI on SU's side
This commit finalizes the use of ROI on Su's side. I add in the
channelTable_entry two members sdp_width and sdp_height in which I save
the resolution parameters got from the SDP file. They are use to
provided security when the user attemps to modify ROI or Channel
resolution parameter.