pipeline.c: correct a Bug in ROI initiation
The ROI initiation for SU is tricky. Indeed, for SUs there is now
videoFormat_entry associated in the videoFormatTable. So we need to
alloc a videoFormat_entry and copy into that the values of the ROI
parameters from the channel when a modification to the ROI parameters is
done. However, during initiation, we were updating the values of the
created videoFormat_entry with initiated values from the channel_entry.
This could not work. I have corrected that.

Another solution would have been to have a different behaviour for
vivoecrop for SP and SU. But I don't think that this is the best way to
do because, anyway the vivoecrop element need the max resolution of the
video to work. For SU this is the SDP's resolution. I prefer to store
that resolution in the channle_entry and copy it to a VF_entry that
giving it to the vivoecrop element in parameters.