channelTable.c: correct error in updateEntry()
It seems to have an error in the function channelTable_updateEntry().
Indeed when attempting to change the channelVideoFormat associated to a
channel nothing happens. So I tried to correct that. I think the problem
is that there are 2 VF enable (with a PLAYING) pipeline that are running
on the same multicast address. The old VF associated to the channel is
still running on the IP address associated to the channel. So there is
also a problem when 2 VF are enable on the same channel: the pipeline
needs to be duplicated, and then the element updsink need to be updated
with the right multicast address corresponding to the channel.