channelTable.c: correct a problem when changing VF
Changing the VF associated to the channel need to be performed only when
the videFormat to change are disable. Otherwise it could mean that the
same videoFormat could be enable on 2 different channel which make no
sense, unless we duplicate it, but as this is multicast, duplicate the
pipeline associated to the video Foramt would only cause a loss of
memory space, but is would not increase the functionnality.

Furthermore, this prolbem makes me corrected another error in
stream-registration.c. The problem was on the registration of a VF. 2 VF
identical could not be added in the VideoFormat_Table. But this is not
true. Indeed, two perfectly identic VF could be 2 cameras that have
different poibnt of view. So I greatly simpplified the function
initialize_videoFormat, which now only creates and add a
videoFormat_entry in the table.