mib-config.c: add key ethernetIpAssignment
This key have been added to the group [deviceInfo]. It will be used so
the user can specify through the configuration file which IP assignment
scheme he wants OpenViveo to use: classic DHCP or system configuration,
or VIVOE IP assignment scheme.
For now I have placed the key's initiation outside the function
init_deviceInfo because I have not put ethernetIpAssignement key in the
deviceInfo table. This could be modified in the future. I made that
choice because this key is not a MIB parameter that will be accessed
later through SNMP request so It might be a good idea to sperate it from
other keys that are MIB objects.