create ip_assignment.c and .h files
This files will contain VIVOE's IP assignment scheme, as described in
the VIVOE Standard. I already added two functions set_static_ip() and
detect_confilct(). The first one assigned a static IP given in
parameters to the interface also given into parameters.
The second one will be use to detect conflict(s) just as it is described
in the MIB.