log.c: replace default logging function
The Glib implements some logging functions. However it commands to the
user to have blank spaces between the critical messages in log. This is
not cool. I don't think that it should impose this to the user. Anyway,
I re-used the code of the default handler ofr the log function but I
removed the blank lines. I will also add the date or the execution time.