conflict_detection.c: send ARP messages
We are now able to send ARP messages. To do so, there are 2 functions
that create the ARP packet: either the PROBE packet or ANOUNCEMENT
packet. Then, a function is used to open a RAW socket, create an layer 2
Ethernet packet, add the ARP message to its payload and then send this
packet to the BROADCAST ethernet address.
It remains to add a function to receive the ARP reply. Then we will need
to check for the conflict and implement the VIVOE IP conflict
resolution.