conflict_detection.c: changes to make ARP work
ARP was not working well. Several reasons to that, first of all there was
a Segmentation Fault in the code. the "eh" variable which is a
ethernet_header structure was used uninitialized in the code. I corrected
that. Then, using Wireshark I saw that the ethernet protcol type was
0x0608 when it should have been 0x0806. This is a typical Endianess issue
with x86 processors that are Little Endian, when the all network is Big
Endian. Well I corrected that, and it seems that I am now able to send
Ethernet packets.