conflict_detection.c: change Endianness of vars
Several things were wrong with the Ethernet frame emmited. First of all
the length of the Ethernet frame was 1024, when we only need a 42 bytes
frame. I corrected that. Then, using wireshark, I noticed that in the
ARP packet the fields "hardware size" and "protocol size" were set to 0.
This was an Endianess problem. I corrected that.
Finally the packet seems fine now: I am indeed sending a "Who has"
packet according to wireshark. However, the target IP address does not
seem to be set to an appropriate value. I will check and eventually
correct that.