conflict_detection.c: make ARP probe work
ARP Probe seems to work just fine now according to the wireshark logs.
The modifications I made to make it work: I change the Target
destination IP, it was set to 10.10.10.10 when it should not. The reason
was that I used the memset() function which only takes an integer in
parameter and set it to the all the memory location. So it was taken
only the first byte of the IP into account. I change it with a memcpy()
so now the memory location that contains the IP address is copy to the
destination memory location.
Another change I made is to compute the Ethernet frame length instead of
just using a #define.