conflict_detection.c: receive ARP reply to request
We were already able to send ARP requests. However, this is a bit useless
if we are not able to to catch the reply to those requests, right? So I
implement a function that opens a RAW socket, for all Ethernet Protocols
and wait for an Ethernet packet which is an ARP reply message. We can
assume that this will be our answer.