ip_assignment.h: replace structure by a global var
The structure ip_tested used to contains a table of IP and the number of
tested IP. However, to be easier to use we will only use a table in
which each index will correspond to the last byte if the tested IP. For
example if the IP 192.168.204.219 has been tested, this IP will be
strored in the row #19. This will give us a rapid look on the IP that
have been tested already.