ip_assignment.c: change randomnnes of IP selection
It is specify in the standard that we should select a random IP between
192.168.204.200 and 192.168.200.253 if a conflict is detected. However
this is not really efficient. Indeed, we can keep a table of tested IP
as I wanted to do but this could lead to a large waste of time when
there is only one IP left. Indeed, if we are selecting IP randomly it
could take some time before hitting the good one. I prefer to start with
the IP 192.168.204.200 and then increase the last byte until we reach
192.168.200.253.