Implement IP conflict detection and resolution scheme
I have now completely implemented the IP assignment scheme of VIVOE. On
startup a device is assigned ti a static IP that he reads from the
configuration file. It performs the IP detection and resolution scheme.
It remains to save the assigned IP addresss for the manager in the non
volatile memory i.e. the configuration file.