ethernetIfTable.c: detect conflict only when no IP assigned
The conflict detection scheme was performed on start up no matter
if there was no IP assigned in the configuration file. Yet, the IP
conflict detection and resolution scheme should not be performed when no
IP are assigned. Indeed, the default IP will be statically used by the
device, waiting to be configured with a valid IP by the manager. And
this is at this time that we should launch the IP detection scheme.