ethernetIfTable.c: implement IP changes from SNMP
The manager can now assigned a new IP to the user if needed. When a new
IP is assigned to the user, this one will perform the IP detection
scheme. However, is the device was not configure to use VIVOE's IP
assignment scheme in the first place, it will reject the request and
return a NOACCESS error.
I also change the meaning of the "index" value in the
ethernetIfTableEntry. Indeed, this was initialize with the index of the
ethernet interface that we get using an ioctl call. However, this could
lead to problems when we are using only one interface. For example, I
configured VIVOE to use only the eth0 interface. This interface have the
index 2 on my machine: 1 is loopback address. So it would has been
intialize with the index 2 in the ethernetIfTable. But the function
implementd by mib2c use the index of the interface and would have never
found the eth0 interface in the table because it is store in first
position and should have the index 1 though.
The VIVOE standard specifies that the index of the entries should be :
"The index for each entry in the table of Ethernet
network interfaces on the device. The index value range
shall be from 1 to the number of entries is specified by
ethernetIfNumber."
I change OpenVivoe to respect that.