mib-config.c: change key name for assigned IP
I thought that the key assignedIP could use what is called "locale"
keys. But it turns out that this is use to have multiple key string in
different languages. This is why only the function get_locale_string()
is implemented.  Finally I just change the key name from assignedIP to
assignedIP_ifname wher ifname is the name of the interface we want to
assigned a new IP for. It seems to work jsut fine now.