ethernetIfTable.c: fix a bug in SET_RESERVE_2
In SET_RESERVE_2 we check if the device is conifgure for the default
system's IP assignment scheme or the VIVOE one. IS it is using VIVOE's
IP assignment scheme, the user is allowed to change the device IP,
otherwise we send a NOACCESS error. However, previously the two cases
were inverted. This has been corrected.