multicast.c:improve computation of multicast IP
I improve the way of compute the multicast IP addresses for the device's
channels. Indeed, these addresses are computing by taking the last byte
of the device IP address. This address was recovered whitin an IOCTL
call. But this does not take advantage of the MIB associated to the
device that already contains the IP address. So I change the code to be
able to compute the multicast IP address from the ethernetIfTableEntry.