mib-config.c: get list of string values for keys
This is an implementation of the function g_key_file_get_string_list()
but with the appropriate checks and error message for the user. I have
used this new function add a function that will retrieve the command line
to use to control physically a camera. This have to be a list of string
because we want to be able to control the camera within the MIB. So the
user will set up a commad line with some %d in it that will be replace
with the values extracted from the MIB. For example the key in
vivoe-mib.conf will be :
camera_ctl="launch-cam top=%d bottom=%d left=%d right=%d"
,top,bottom,left, rigth

Then the list top,bottom,left, rigth just gives theorder of the
parameters in the commandline, and will then be replaced with their
conrresponding values, extracted from the MIB.