mib-config.c: change key camera_ctl type
This key was a list of string. It used to contain a command line to
control the camera, whith some '%d' characters. Then follows the
variables from the MIB that correspond to the '%d' and use to replace
them. I have change that so the key camera_ctl contains only a string
but with some characters %a %b %c %d in it. Then %a will be replace with
the value of channelRoiOriginTop, %b will be replaced by the value of
channelOriginLeft etc.