roi.c: control camera with the MIB
There was only one type of ROI in OpenVivoe. To add a ROI in a pipeline
you needed to use specific plugins called vivoecrop and vivoecaps. Now
the possibility is also given to the user to control a camera within the
MIB. To do so he must fill a key in the configuration file named
"camera_ctl" with a command line that will be use to control tge camera.
I have add two functions in roi.c, one which will be use replace the
value %a %b %c %d in the command line with the values of
channelRoiOriginTop, channelRoiOriginLeft, channelRoiExtentBottom and
channelRoiExtentRight. The second function executes the command in a new
shell and exit the shell.