pipelince.c: mv handle_roi in addSink_SU
This was preformed after the sink is added in the pipeline, however we
may want to add a branch in this pipeline in order to support
redirection with sink. So I movz it before the sink is linked to the
previous element in pipeline so we can create a branch with appsink and
the typefind element.