channelTable.c: correct a SegFault
When a ROI change fail. One of the reason could be that the user tries
to change a ROI parameter on a Service User or Service Provider that
does not support ROI. Then channelTable_hanlder returns FALSE and we go
to set undo where we remove the changes made to the videoFormat
associated to the channel before the ROI changed. However, if it is in
case of a Service User, there is no videoFormat associated to the
channelEntry. Thus, when it means that the handler will try to undo
changes never done. This ended with a SegFault. I corrected that.