Implement Redirection ROI
I have implemented the redirection with ROI support. It seems to work
fine. To implement it I have create two branches in the pipelines of the
SU and the SP:
_ In SU pipeline I created a branch just before the appsink
_ In Sp pipeline I created a branch just before the udpsink (as normal
ROI)

Then when a ROI changes is detected for a redirection. We detect new
caps before appsink, give them to appsrc, and re-compute the caps before
udpsionk to update the channel Entry and have a SDP file matching the
caps.

It seems to work but as my computer is really slow, even for a normal
redirection, I cannot be 100% sure.