pipeline.c: return if bin sink not build
On the Service User side the when the sink could not be built from
configuration file, we try to set it a name even though. This has been
corrected.
Also in roi.c I moved the detection of the redirection before te
detection of the video Format so we don't update caps twice if this is a
ROI redirection of an mpeg4 video.