pipeline.c:improve restriction for J2K in SU redirect
Rtpj2kdepay and rtpj2kpay only support one kinf of J2K: image/x-jpc.
The first solution implemented was to add a capsfilter. But the caps of
this capsfilter was built by hand, using the string "image/x-jpc"
however, this is not the best way to do. The best way to do is what has
been implemented now. The caps given to the capsfilter are built from
the sink pad of rtpj2kpay only, which is much more interresting because
it will also handle the new kind of J2k when rtpj2kpay will support
them.