set primary interface as default for multicast
The multicast did not work on local network. The reason was that the
device could not find a proper route to send multicast packets. This
has been changed so the device select the primary interface, i.e.
the first node in the ethernetIfTable, as deafult interface to send its
packets. This has been changed for both the SAP/SDP socket, and for the
updsrc/udpsink element.