add a way to build Documentation with cmake
Cmake is now also able to provide a way of building the project
Documentation thanks to Doxygen. To do so, I update the CMakeList.txt
for it to check for the Doxygen package first, and if it is present it
will add a PHOENIX rule "doc" to make that will build the documentation.
The documentation will be built in the doc folder.