mib-config;c: exit if mandatory key not found
Before, if a mandatory key was not found, the program did not exit. This
lead to major failure, as for example if the key which give the name of
the ethernetInterface is not found in the configuration file.
Besides, I have replace some functions which used to return int value
for them to return a boolean, which is clearer for everybody.