ethernetIfTable.c: don't get interface index
The Interface Index field in the ethernetTable of the group deviceInfo
of the VIVOE-MIB does not have the same meaning as the name of the real
Interface Index. The interface index we were getting with an IOCTL call
is the interface's index on the device. We will get this parameter only
when we will perform the IP conflict detection and resolution.
Otherwise, this value is not usefull.