conflict_detection.c: make the conflict detection work
The conflict detection used to work only on dhcp network but not in
local network. The reason was that the route for ARP packets was not set
properly. I have corrected that. To send level 2 ethernet packets, the
kernel ask to the user to fill a struct called socket_addr with the
interface index of the interface to use to send the packet. This is how
the user builds the route for the packet. The problem here was that I
gave the ethernetIfEntry_index value which is the index of the interface
in the ethernetIfTable but not the index of the interface on the device.
So OpenVivoe used to send the packet with loopback interface instead of
the classic interface.
Then I updated the function so that it open the sending socket only once
and build the Probe message only once. This is way more efficient as we
send 3 times the same packet. Then, each time a conflict is detected we
rebuild a new ethernet frame, that we sent 3 times, through the same
socket. At the end of the ip conflict detection the ethernet frame is
freed and the socket is closed.