source: make proper #includes
Includes were not done properly. Instead I used to compile with the
relative path "." in my gcc with the option -I. This is a not a good
practice. Instead I modified my CMakeList.txt to remove this gcc option,
and I modified the sources. Instead of including files in my *.c as
\#include "include/*.h" it is no
\#include "*.h"

This will let the user to build the source outside the project directory
as it is done in yocto.