CMakeList.txt: check for define types
There was some errors during building on some machines because in
openvivoe, some of the primitive C types such as u_long were not
defined. I have modified the CMakeList.txt so it now checks if the
headers in which those types are defined, are indeed included in the
code. If so, it creates in the "include" folder a file named config.h
that define the appropriate macros to defind those types.